+++ CHANGELOG.md
@@ -25,6 +25,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Number of records in the pipeline table in the web
   console is no longer rounded
   ([#967](https://github.com/feldera/feldera/pull/967))
+- Ability to pause/resume row updates in Data Inspection tables
+  ([#603](https://github.com/feldera/feldera/issues/603))
+
 
 ### Fixed
 - Reduce Docker logging noise from Kafka connect and Redpanda.